backend/requirements.txt
Django>=3.2,<4.2
djangorestframework>=3.12,<4.0
psycopg2-binary  # For PostgreSQL, use mysqlclient for MySQL
djangorestframework-simplejwt  # For JWT authentication
django-cors-headers  # To handle CORS (Cross-Origin Resource Sharing)
djangorestframework  # Optional: for API documentation
pytest  # For testing
pytest-django  # For Django testing support